Creates TexShip class
TexShip is meant to be a class for the player's ship. Might change to be Ship.h & Ship.cpp in the future.
TexShip inherits from TexRect but will soon hold functions that will be more specific to the class (like movement and shooting functions).
contains() was added to act as the rough hit box of the player ship. It returns true or false based on whether a coordinate is within the boundaries of a triangle, rather than a rectangle.
moveShip() is meant to be the function dedicated to moving the ship across the x axis. Will be implemented later.
All other edits were minor with no real influence yet.