Allows ship to move
The ship can now move left and right. Though the movement is choppy. Will look into fixing that later.
Renamed the TexShip class to Ship.
Added code to check the hit box of the ship, by clicking the ship and displaying a message in terminal to indicate that the ship has been hit (or in this case, clicked). This is merely for testing purposes and will be modified later to work with bullets.