Creates smooth motion for player ship
Utilizes idle() function to make smooth motion of the ship. Since the idle() function has a higher refresh rate than the keyDown() function, it makes the ship move at a higher refresh rate. Two global boolean variables were created to keep track of which buttons were pressed.
Ship.cpp was edited to simply decrease the amount that the ship will move when a key is pressed, so the ship doesn't move too fast.