Adds aliens
Created a Minion class that inherits from the TexRect class. Might change this to have an Alien class inherit from TexRect and then have a Minion and MegaAlien inherit from Alien to allow them to both have common methods between them.
Adds functionality to control the player ship with the arrow keys.
The aliens deque in Game.cpp has been modified to hold Minion pointers for simplicity. This should be changed in the future.
Edited the renderText() method in the Text class to allow for the input of rgb values to render the text a certain color. This is good because it allows there to only be one Text pointer rather than two, as before.
The draw() function was deleted from the Ship class, after realizing it was completely worthless since the class could simply inherit the draw() function from the parent class TexRect.
Some images were deleted due to lack of time to be able to use them.
Edited app_main.cpp to simply change the window title.