Puts both aliens under Alien.h
Creates an Alien class to unify both the Minion and MegaAlien classes. This will be used to give them common methods such as shoot() which will allow the aliens to shoot a bullet at the player.
Also creates the Bullet class which inherits from TexRect. This class will be used as a parent class to ShipBullet and AlienBullet, since both should have fairly similar functionality.
Game was modified to have both types of aliens in the aliens deque, thus drawing them on the window.