Gives ability to shoot
Gives the ship the ability to shoot bullets. Bullets still do not interact with aliens though.
Shooting is enabled via the space bar.
Edited the Bullet constructor to have a better system for the (x,y) coordinates. Coordinates put into the constructor, now represent the tip of the bullet, rather than the top left corner of the image.
Modified Bullet.move() to move a set amount in the y axis, however the method was also changed to be virtual, meaning if AlienBullet gets created, AlienBullet can have a different move() function that moves down (towards the player) rather than up.